Further breakup player mod
